This is a major re-architecture of how Proof-of-Space is implemented.

It sacrifices the completeness of the tables for high-performance parallel processing, which is already faster on the CPU and will be very beneficial for GPU implementation as well. See subspace diff document for brief details, I'll also write a blog post describing the changed in more detail.

As the result, table creation is substantially faster, there are less allocations, memory usage has also decreased substantially, probably up to 80%.

The previous best observed result for parallel table construction was:
```
chia/table/parallel/8x  time:   [677.60 ms 684.25 ms 692.06 ms]
                        thrpt:  [11.560  elem/s 11.692  elem/s 11.806  elem/s]
```
With this PR performance is much higher:
```
chia/table/parallel/8x  time:   [543.21 ms 549.05 ms 553.93 ms]
                        thrpt:  [14.442  elem/s 14.571  elem/s 14.727  elem/s]
```

I saw 10ms lower times with slightly different implementation that instead of collecting results into final vectors right away was instead collecting results into intermediate combined vector. While it was a tiny bit faster and simplified the code in a bunch of places, it required more memory in the process. I think now that most places don't use `ys` as a list and values are scattered in RAM anyway, we can combine everything together for both sequential and parallel variant (probably except metadata, which is now freed to reduce peak memory usage and allow allocator to reuse memory), but that will be an exercise for another day.

Single-threaded table creation also improved massively:
```
Before:
chia/table/single/1x    time:   [920.37 ms 924.24 ms 929.24 ms]
                        thrpt:  [1.0762  elem/s 1.0820  elem/s 1.0865  elem/s]
After:
chia/table/single/1x    time:   [750.41 ms 760.79 ms 772.80 ms]
                        thrpt:  [1.2940  elem/s 1.3144  elem/s 1.3326  elem/s]
```

Proof generation is slightly slower, but considering up to $$2^{16}$$ will be queried, the tiny increase in time (~4ms for each encoded piece that will likely be amortized anyway) is certainly worth the performance increase. Also majority of the slowdown is for missing proofs, which are the minority.
```
Before:
chia/proof/missing      time:   [20.764 ns 21.179 ns 21.459 ns]
                        thrpt:  [46.600 Melem/s 47.217 Melem/s 48.160 Melem/s]
chia/proof/present      time:   [360.24 ns 360.65 ns 361.08 ns]
                        thrpt:  [2.7695 Melem/s 2.7727 Melem/s 2.7760 Melem/s]
After:
chia/proof/missing      time:   [101.94 ns 102.22 ns 102.58 ns]
                        thrpt:  [9.7486 Melem/s 9.7824 Melem/s 9.8099 Melem/s]
chia/proof/present      time:   [376.64 ns 377.52 ns 379.55 ns]
                        thrpt:  [2.6347 Melem/s 2.6489 Melem/s 2.6551 Melem/s]
```

There are still a few optimizations to explore with new architecture and, as before, BLAKE3 SIMD will also provide a substantial boost here.

Very happy with the result after countless fruitless attempts and unsuccessful (performance-wise) rewrites, including alternative search implementations and other tweaks. GPU implementation will follow the same general pattern and should be implementable a lot more efficiently than the previous algorithm (while remaining compatible with this updated CPU implementation).

I considered having two incompatible implementations: one CPU-optimized (that can't be used with GPU) and another GPU-optimized (that will have a CPU version as a fallback), but for now I think we can get away with this version.

There are also some optimizations that would introduce non-determinism, which could still be exploited if a trace is recorded about the exact path the code followed during table creation, essentially trading faster initial table creation for slower reconstruction afterward. Don't think it is worth the complexity either, but it is a hypothetical possibility.

So many options to consider after thinking about it during last few weeks!